+++ src/api/stories/stories.ts
@@ -7,14 +7,26 @@ import type {
     UpdateStory,
     UpdateStories,
     RemoveAllStories,
+    UpsertStory,
+    DeepUpsertStory,
+    ExtendedISbStoriesParams,
+    GetStoryBySlug,
 } from "./stories.types.js";
-import type { GetStoryBySlug } from "./stories.types.js";
 
 import chalk from "chalk";
 
-import storyblokConfig from "../../config/config.js";
 import Logger from "../../utils/logger.js";
+import { managementApi } from "../managementApi.js";
 import { getAllItemsWithPagination } from "../utils/request.js";
+
+const notNullish = <T extends Record<string, any>>(params: T): T => {
+    return Object.keys(params).reduce((acc, key) => {
+        if (params[key] !== null && params[key] !== undefined) {
+            acc[key] = params[key];
+        }
+        return acc;
+    }, {} as any);
+};
 
 export const removeStory: RemoveStory = (args, config) => {
     const { storyId } = args;
@@ -38,7 +50,7 @@ export const removeAllStories: RemoveAllStories = async (config) => {
     Logger.warning(
         `Trying to remove all stories from space with spaceId: ${spaceId}`
     );
-    const stories = await getAllStories(config);
+    const stories = await getAllStories({}, config);
 
     const onlyRootStories = (story: any) =>
         story.story.parent_id === 0 || story.story.parent_id === null;
@@ -56,16 +68,27 @@ export const removeAllStories: RemoveAllStories = async (config) => {
 };
 
 // GET
-export const getAllStories: GetAllStories = async (config) => {
+export const getAllStories: GetAllStories = async (args, config) => {
+    const { options } = args;
     const { spaceId, sbApi } = config;
     Logger.log(`Trying to get all Stories from: ${spaceId}`);
+
+    const params = notNullish<ExtendedISbStoriesParams>({
+        with_slug: options?.with_slug,
+        starts_with: options?.starts_with,
+        language: options?.language,
+    });
+
+    console.log("These are params i will use: ");
+    console.log(params);
 
     const allStoriesWithoutContent = await getAllItemsWithPagination({
         apiFn: ({ per_page, page }) =>
             sbApi.get(`spaces/${spaceId}/stories/`, {
+                ...params,
                 per_page,
                 page,
-            }),
+            } as ExtendedISbStoriesParams),
         params: {
             spaceId,
         },
@@ -101,8 +124,8 @@ export const getAllStories: GetAllStories = async (config) => {
 
 // GET
 export const getStoryById: GetStoryById = (storyId, config) => {
-    const { spaceId, sbApi } = config;
-    if (storyblokConfig.debug) {
+    const { spaceId, sbApi, debug } = config;
+    if (debug) {
         console.log(
             `Trying to get Story with id: ${storyId} from space: ${spaceId}, to fill content field.`
         );
@@ -110,7 +133,7 @@ export const getStoryById: GetStoryById = (storyId, config) => {
     return sbApi
         .get(`spaces/${spaceId}/stories/${storyId}`)
         .then((res: any) => {
-            if (storyblokConfig.debug) {
+            if (debug) {
                 Logger.success(
                     `Successfuly fetched story with content, with id: ${storyId} from space: ${spaceId}.`
                 );
@@ -163,10 +186,15 @@ export const updateStory: UpdateStory = (content, storyId, options, config) => {
     Logger.log(
         `Updating story with name: ${content.name} in space: ${spaceId}`
     );
+
+    console.log("THis is content to update: ");
+    console.log(JSON.stringify(content, null, 2));
+
     return sbApi
         .put(`spaces/${spaceId}/stories/${storyId}`, {
             story: content,
             publish: options.publish ? 1 : 0,
+            force_update: options.force_update ? 1 : 0,
         })
         .then((res: any) => {
             console.log(`${chalk.green(res.data.story.full_slug)} updated.`);
@@ -176,7 +204,6 @@ export const updateStory: UpdateStory = (content, storyId, options, config) => {
 };
 
 export const updateStories: UpdateStories = (args, config) => {
-    const { sbApi } = config;
     const { stories, options, spaceId } = args;
 
     return Promise.allSettled(
@@ -186,8 +213,88 @@ export const updateStories: UpdateStories = (args, config) => {
                 stories.story,
                 stories.story.id,
                 { publish: options.publish },
-                { sbApi, spaceId }
+                { ...config, spaceId }
             );
         })
     );
+};
+
+export const upsertStory: UpsertStory = async (args, config) => {
+    console.log("Modifying story... in space with id:");
+    console.log(config.spaceId);
+    const { storyId, storySlug, content } = args;
+
+    console.log("This are args passed: ");
+    console.log(args);
+
+    if (storyId) {
+        // if this exist than we update story with this id
+        console.log("You've selected storyid!");
+    } else if (storySlug) {
+        // if this exist than we update story with this slug (probably when we try to add story from one space to another,
+        console.log("You've selected slug!");
+        const foundStory = await managementApi.stories.getStoryBySlug(
+            storySlug,
+            config
+        );
+        console.log("This is story");
+        console.log(foundStory);
+
+        if (foundStory) {
+            // then update the story
+        } else {
+            const {
+                story: { parent_id, id, parent, ...rest },
+            } = content;
+            console.log("We are going to create story");
+            const response = await managementApi.stories.createStory(
+                rest,
+                config
+            );
+            console.log("This is response");
+            console.log(response);
+        }
+    } else {
+        // if this exist than we create new story
+        console.log("Nothing passed, creating story...");
+    }
+};
+
+export const deepUpsertStory: DeepUpsertStory = async (args, config) => {
+    console.log("Modifying story... in space with id:");
+    console.log(config.spaceId);
+    const { storyId, storySlug, content } = args;
+
+    console.log("This are args passed: ");
+    console.log(args);
+
+    if (storyId) {
+        // if this exist than we update story with this id
+        console.log("You've selected storyid!");
+    } else if (storySlug) {
+        // if this exist than we update story with this slug (probably when we try to add story from one space to another,
+        console.log("You've selected slug!");
+        const slugs = storySlug.split("/");
+        console.log(
+            "Slugs for which we need to check for existence of stories"
+        );
+        console.log(slugs);
+
+        // const foundStory = await managementApi.stories.getStoryBySlug(storySlug, config)
+        // console.log("This is story")
+        // console.log(foundStory)
+
+        // if(foundStory) {
+        //     // then update the story
+        // } else {
+        //     const {story: {parent_id, id, parent,...rest}} = content
+        //     console.log("We are going to create story")
+        //     const response = await managementApi.stories.createStory(rest, config)
+        //     console.log("This is response")
+        //     console.log(response)
+        // }
+    } else {
+        // if this exist than we create new story
+        console.log("Nothing passed, creating story...");
+    }
 };